# api/openapi/payments-platform.v1.yaml
openapi: 3.1.0
...
info:
  title: Payments Platform API
  version: 1.0.0
  summary: Multi-tenant payment gateway onboarding and payment orchestration.
  description: |
    The public REST surface of the payments platform. Two products share this contract:

    * **Control plane** (`/v1/merchants`, `/v1/gateways`) — merchant registry, onboarding
      workflow, desired-state configuration and the gateway catalogue. Availability target
      99.9 %. Never on the payment hot path.
    * **Data plane** (`/v1/payments`, `/v1/webhooks`) — payment orchestration against
      third-party gateways. Availability target 99.99 %, p99 <= 250 ms excluding gateway time.

    ## Binding rules

    * **Money** is always an integer in the currency's minor units plus an ISO-4217 code:
      `{"amount": 1050, "currency": "USD"}`. Never a float, never a decimal string.
    * **Identifiers** are `<prefix>_<26-char Crockford Base32 ULID>` and are opaque to
      clients. Do not parse them.
    * **No cardholder data.** This API accepts gateway tokens and vault references only.
      Any request field containing a Luhn-valid 13-19 digit sequence is rejected with
      `400 SENSITIVE_DATA_IN_REQUEST` and the value is never logged. See
      `PaymentMethodReference`.
    * **Idempotency.** Every unsafe operation requires an `Idempotency-Key` header. Replays
      return the stored response with `Idempotent-Replay: true`. A key reused with a
...
      `If-Match`. A mismatch is `412`.
...
  contact:
...
  license:
...
servers:
...
security:
  - oauth2:
      - merchants:read
paths:
  /v1/merchants:
    post:
      operationId: createMerchant
      tags: [Merchants]
      summary: Create a merchant
      description: |
        Registers a merchant under the authenticated principal's tenant in state `CREATED`
        and publishes `merchant.created.v1`. The tenant is taken from the access token; a
        `tenantId` supplied in the body is rejected with `403 TENANT_MISMATCH`.

        Creating a merchant does not start onboarding. Call
        `POST /v1/merchants/{merchantId}/onboarding` for that.
      security:
        - oauth2: [merchants:write]
        - mtls: []
      x-idempotent: true
      x-rate-limit:
        scope: tenant
        limit: 60
        window: 1m
        burst: 120
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/XCorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMerchantRequest'
      responses:
        '201':
          description: Merchant created.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Location: { $ref: '#/components/headers/Location' }
            Idempotent-Replay: { $ref: '#/components/headers/IdempotentReplay' }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Merchant'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    get:
      operationId: listMerchants
      tags: [Merchants]
      summary: List merchants
      description: |
        Cursor-paginated list of the tenant's merchants, newest first by ULID order. There is
        no offset pagination: offsets are unstable under concurrent writes.
      security:
        - oauth2: [merchants:read]
        - mtls: []
      x-idempotent: true
      x-rate-limit:
        scope: tenant
        limit: 120
        window: 1m
        burst: 200
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/XCorrelationId'
        - name: status
          in: query
          required: false
          description: Filter by merchant lifecycle state. Repeatable.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/MerchantStatus'
          style: form
          explode: true
          example: [ACTIVE, SUSPENDED]
        - name: externalReference
          in: query
          required: false
          description: Filter by the tenant's own identifier for the merchant. Exact match.
          schema:
            type: string
            maxLength: 128
          example: acme-eu-01
      responses:
        '200':
          description: A page of merchants.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantPage'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /v1/merchants/{merchantId}/configuration:
    parameters:
      - $ref: '#/components/parameters/MerchantIdPath'
    get:
      operationId: getMerchantConfiguration
      tags: [Configuration]
      summary: Get the active configuration
      description: |
        Returns the merchant's currently active desired-state configuration document and its
        version metadata. The `ETag` is the document digest and must be echoed in `If-Match`
        on `PUT`.
      security:
        - oauth2: [config:read]
        - mtls: []
      x-idempotent: true
      x-rate-limit:
        scope: merchant
        limit: 300
        window: 1m
        burst: 600
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: The active configuration document.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationVersion'
        '304':
          description: Not modified.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    put:
      operationId: putMerchantConfiguration
      tags: [Configuration]
      summary: Publish a new configuration version
      description: |
        Replaces the merchant's desired-state configuration. The document is validated (L4),
        assigned the next dense version, persisted with the full prior document retained,
        audited with actor and diff, and published as `configuration.published.v1`.

        Configuration history is strictly append-only: this endpoint never mutates an existing
        version. Propagation to the data plane is asynchronous with a p99 budget of 30 s.

        Requires `If-Match` with the current version's `ETag`. A concurrent publish loses with
        `412` and must re-read, re-merge and retry - the platform does not merge configuration
        documents on the caller's behalf.
      security:
        - oauth2: [config:write]
        - mtls: []
      x-idempotent: true
      x-rate-limit:
        scope: merchant
        limit: 30
        window: 1m
        burst: 60
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/XCorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantConfiguration'
      responses:
        '200':
          description: A new configuration version was published.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Idempotent-Replay: { $ref: '#/components/headers/IdempotentReplay' }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationVersion'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/merchants/{merchantId}/configuration/versions:
    parameters:
      - $ref: '#/components/parameters/MerchantIdPath'
    get:
      operationId: listConfigurationVersions
      tags: [Configuration]
      summary: List configuration versions
      description: |
        Cursor-paginated version history, newest first. Every version - including those
        created by a rollback - is present; nothing is ever deleted or rewritten.
      security:
        - oauth2: [config:read]
        - mtls: []
      x-idempotent: true
      x-rate-limit:
        scope: merchant
        limit: 120
        window: 1m
        burst: 200
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/XCorrelationId'
      responses:
        '200':
          description: A page of configuration versions.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationVersionPage'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/merchants/{merchantId}/configuration/rollback:
    parameters:
      - $ref: '#/components/parameters/MerchantIdPath'
    post:
      operationId: rollbackConfiguration
      tags: [Configuration]
      summary: Roll back to a previous configuration version
      description: |
        Re-publishes the document of `toVersion` **as a new version** `n+1`. A rollback is an
        append, never a deletion, so the audit trail retains the version being abandoned.

        The target document is **re-validated against L4 at rollback time**, not trusted
        because it was valid when it was first published: a gateway may since have been
        de-registered, a certification expired or a residency policy tightened. If
        re-validation fails the response is `422 CONFIGURATION_INVALID` naming the rules that
        now fail. A rollback that cannot be validated is not a safe rollback.

        Emits both `configuration.rolled_back.v1` (the operator signal) and
        `configuration.published.v1` (what the data plane's ordinary handler consumes).
      security:
        - oauth2: [config:rollback]
        - mtls: []
      x-idempotent: true
      x-rate-limit:
        scope: merchant
        limit: 10
        window: 1m
        burst: 10
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/XCorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigurationRollbackRequest'
      responses:
        '201':
          description: |
            A new version was created carrying the restored document. Its digest equals the
            restored version's digest, which is the machine-checkable proof that the rollback
            is faithful.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Location: { $ref: '#/components/headers/Location' }
            Idempotent-Replay: { $ref: '#/components/headers/IdempotentReplay' }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationVersion'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        OAuth 2.0 client credentials. Tokens are `ES256` JWTs with a 15-minute TTL and
        audience `https://api.example.com`.

        **The tenant is taken from the token, never from the request.** A `tenantId` in a body
        or query string that disagrees with the token is a security event:
        `403 TENANT_MISMATCH`, audited and alerted.

        The token endpoint is rate limited to 100 requests / 5 min per IP and 20 / min per
        client. Scopes are minted per client, not requested per call: asking for a scope the
        client was not granted returns `invalid_scope` rather than a narrowed token.
      flows:
        clientCredentials:
          tokenUrl: https://auth.example.com/oauth2/token
          refreshUrl: https://auth.example.com/oauth2/token
          scopes:
            merchants:read: Read merchants, business profiles, bank accounts and principals.
            merchants:write: Create and update merchants.
            merchants:suspend: Suspend and resume a merchant.
            merchants:terminate: Terminate a merchant. Dual control required.
            onboarding:read: Read onboarding cases and step history.
            onboarding:write: Start and drive onboarding.
            onboarding:approve: Signal a manual onboarding gate. Dual control required.
            config:read: Read configuration documents and version history.
            config:write: Publish a new configuration version.
            config:rollback: Roll back configuration to an earlier version. Dual control required.
            gateways:read: Read the gateway catalogue and health.
            gateways:write: Register and deprecate gateways. Platform administrators only.
            credentials:rotate: Rotate gateway credentials. Dual control required.
            credentials:read: Read credential metadata for the caller's own path. Internal workloads only.
            payments:read: Read payments, attempts and refunds.
            payments:write: Create payments. Machine clients only - no human role holds this scope.
            payments:capture: Capture an authorized payment.
            payments:refund: Refund a captured payment. Dual control for human principals.
            payments:void: Void an authorization.
            payments:replay_dlq: Replay dead-lettered payment events. Operators only.
            ledger:read: Read ledger accounts and entries.
            ledger:write: Append ledger entries. Event consumer only.
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-supplied opaque key scoping one logical operation to one effect. Required on
        every unsafe operation.

        The idempotency scope is
        `(tenant_id, merchant_id, method, path_template, idempotency_key)`. A SHA-256
        fingerprint of the JCS-canonicalized request body is stored with the key: the same key
        with a different body is `422 IDEMPOTENCY_KEY_REUSED`, which is the check that catches
        a client reusing one key for two different payments.

        A concurrent duplicate while the first request is still in flight is
        `409 IDEMPOTENT_REQUEST_IN_PROGRESS` with `Retry-After: 1`. The platform does not
        block the second caller and does not process twice.

        Recommended: a UUIDv4 per logical operation, generated **before** the first attempt and
        reused across every retry of that operation. Records are retained 7 days.
      schema:
        type: string
        minLength: 1
        maxLength: 255
        pattern: '^[\x21-\x7E]{1,255}$'
      example: 6b3f1a52-9e7c-4b4f-8f4a-1c2d3e4f5a6b
...
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: |
        Caller-supplied request identifier, echoed on every response including errors. If
        absent the platform generates a `req_`-prefixed ULID. It appears in the problem
        document as `requestId` and on every log line and audit record for this request.
      schema:
        type: string
        maxLength: 128
      example: req_01JB8Z22222222222222222222
    Traceparent:
      name: traceparent
      in: header
      required: false
      description: |
        W3C Trace Context. Propagated into every span, event envelope and gateway call, and
        echoed on the response. If absent a new trace is started. Format
        `00-<32 hex trace-id>-<16 hex span-id>-<2 hex flags>`.
      schema:
        type: string
        pattern: '^[0-9a-f]{2}-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$'
      example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
    XCorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: |
        Business correlation identifier that stays constant across an entire causal fan-out,
        so one API call's full consequences - payments, attempts, events, audit records - are
        one query. Copied into `correlationid` on every event this request produces. Defaults
        to the request id.
      schema:
        type: string
        maxLength: 128
      example: req_01JB8Z22222222222222222222
...